Set methods to non-static, add logging, extract common unit test code
