+++ src/test/java/net/wasdev/wlp/common/plugins/util/InstallFeatureUtilGetServerFeaturesTest.java
@@ -1,3 +1,18 @@
+/**
+ * (C) Copyright IBM Corporation 2018.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package net.wasdev.wlp.common.plugins.util;
 import static org.junit.Assert.assertEquals;
 
@@ -13,9 +28,10 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-public class InstallFeatureUtilGetServerFeaturesTest {
+public class InstallFeatureUtilGetServerFeaturesTest extends BaseInstallFeatureUtilTest {
     private static File serverDirectory = null;
     private static File src = null;
+    private static InstallFeatureUtil util = null;
 
     @Rule
     public TemporaryFolder tempFolder = new TemporaryFolder();
@@ -28,6 +44,7 @@ public class InstallFeatureUtilGetServerFeaturesTest {
     @Before
     public void setUp() throws Exception {
         serverDirectory = tempFolder.newFolder();
+        util = getNewInstallFeatureUtil();
     }
 
     private void copyAsName(String origName, String newName) throws IOException {
@@ -40,8 +57,8 @@ public class InstallFeatureUtilGetServerFeaturesTest {
         FileUtils.copyFileToDirectory(file, serverDirectory);
     }
     
-    private void verifyServerFeatures(Set<String> expected) {
-        Set<String> getServerResult = InstallFeatureUtil.getServerFeatures(serverDirectory);
+    private void verifyServerFeatures(Set<String> expected) throws Exception {
+        Set<String> getServerResult = util.getServerFeatures(serverDirectory);
         assertEquals("The features returned from getServerFeatures do not equal the expected features.", expected, getServerResult);
     }
